test(resolver): move the term-unavailable regression to resolver_test
The wasip1 term-loading regression test landed in pkg/bytecode's
lgb_integration_test.go because that file already had the compiler+resolver+rt
setup it needs. It exercises resolver namespace-loading behavior, so
pkg/resolver/resolver_test.go is the discoverable home (per review nit). No
import cycle: resolver already imports compiler/rt/vm, so an internal-package
test can too. Behavior unchanged.
